# src/main/java/com/apromore/challenge/dto/CourseResponseDTO.java
package com.apromore.challenge.dto;

import lombok.*;

@Getter
@Setter
@ToString
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class CourseResponseDTO {
    private Long id;
    private String name;
    private String code;
}